backend/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["bani2016up <bani2016up@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
supabase = "^2.15.0"
python-jose = "^3.4.0"
pydantic = "^2.11.3"
dotenv = "^0.9.9"
uvicorn = "^0.34.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
